Split chart stories into several js modules
